requirements.txt
yfinance>=0.2.18
praw>=7.7.1
requests>=2.31.0
pandas>=2.0.3
numpy>=1.24.0
python-dotenv>=1.0.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.15.0
jupyter>=1.0.0